Tweaking code to make it more performant, as per the advice of @EdgeCaseBerg. Changes include:
- removing the len(subgroup) == k comparison
this is unnecessary if we instead loop through the range of
only the items in the list - the number of items from the
list we want to select.

- removing the slices into the list to create a subgroup
This is an unnecessary and expensive operation which can be
replaced by simply accessing the appropriate max and min items
by supplying the index directly.