Method of selecting list elements consistently
This defines a function that, when called with a list of elements, the number of shards to break the list into, and the shard of interest, will return the same set of elements each time. This is useful when sharding some list of elements into multiple pieces, when you desire to select the same piece each time. E.g., with numshards=2, the list ['a', 'b', 'c', 'd'] will be broken into ['a', 'b'] and ['c', 'd'], and if shard index is set to 1, the sublist ['c', 'd'] will always be returned.